% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdw.R
\name{clean_sdw_cache}
\alias{clean_sdw_cache}
\title{Clean SDW Cache}
\usage{
clean_sdw_cache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{A path to cache directory. If \code{NULL} (default) tries to clean the default temporary cache directory.}
}
\description{
Delete the sdw cache directory.
}
\examples{
\dontrun{
clean_sdw_cache()
}
}
\seealso{
\code{\link[=get_sdw]{get_sdw()}}
}
