% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{close_conn}
\alias{close_conn}
\title{Closes an ODBC connection}
\usage{
close_conn(conn)
}
\arguments{
\item{conn}{An ODBC connection object}
}
\description{
Closes an ODBC connection
}
\seealso{
\code{\link[=get_conn]{get_conn()}}
}
