% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{get_conn}
\alias{get_conn}
\title{Creates an ODBC connection to an SQL Server instance}
\usage{
get_conn(
  server,
  database = NULL,
  timezone = "Europe/Helsinki",
  timezone_out = "Europe/Helsinki"
)
}
\arguments{
\item{server}{SQL server name}

\item{database}{Database name (optional)}

\item{timezone}{The Server time zone. Default is 'Europe/Helsinki'. See \code{\link[odbc:dbConnect-OdbcDriver-method]{odbc::dbConnect()}}}

\item{timezone_out}{The time zone returned to R. Default is 'Europe/Helsinki'. See \code{\link[odbc:dbConnect-OdbcDriver-method]{odbc::dbConnect()}}}
}
\description{
Creates an ODBC connection to an SQL Server instance
}
\seealso{
\code{\link[=close_conn]{close_conn()}}
}
