% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdw.R
\name{get_sdw}
\alias{get_sdw}
\title{Get data from an SDW dataset}
\usage{
get_sdw(
  dataset,
  key,
  parse_dates = TRUE,
  start = NULL,
  end = NULL,
  first_n_obs = NULL,
  last_n_obs = NULL,
  confidential = FALSE,
  auth_token_id = "",
  username = NULL,
  password = NULL,
  verbose = FALSE,
  use_cache = TRUE,
  cache_dir = NULL,
  refresh_cache = TRUE,
  update_cache = FALSE
)
}
\arguments{
\item{dataset}{Name of the queried SDW dataset}

\item{key}{An SDMX key string (without the dataset) or a list of individual SDMX key items}

\item{parse_dates}{Whether to cast TIME_PERIOD strings to R dates}

\item{start}{Starting time period}

\item{end}{Ending time period}

\item{first_n_obs}{Number of observations from starting period onwards}

\item{last_n_obs}{Number of observations from ending period backwards}

\item{confidential}{Whether to use authentication}

\item{auth_token_id}{Windows keyring ID with the authentication information}

\item{username}{Alternative username to use for authenticating}

\item{password}{Alternative password to use for authenticating}

\item{verbose}{Whether the print request debug information}

\item{use_cache}{Whether to use caching}

\item{cache_dir}{Directory for storing cache files}

\item{refresh_cache}{Whether to enable automatically updating the cache}

\item{update_cache}{Whether to force updating the cache}
}
\value{
A dataframe containing the data from the SDW request
}
\description{
Get data from an SDW dataset
}
\examples{
get_sdw("BSI", key = list("M", "", "N", "A", "T00", "A", "1", "Z5", "0000", "Z01", "E"))
get_sdw("BSI", key = list("M.U2.N.A.T00.A.1.Z5.0000.Z01.E"))
}
\seealso{
\code{\link[=clean_sdw_cache]{clean_sdw_cache()}}
}
