% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.loaders-package.R
\docType{package}
\name{r.loaders-package}
\alias{r.loaders}
\alias{r.loaders-package}
\title{r.loaders: Set of Tools for Loading Data From the ECB Data Portal and SQL servers}
\description{
Provides a set of functions to easily get data from the ECB Data Portal and SQL servers
}
\author{
\strong{Maintainer}: Arturs Janis Nikitins \email{ajnikitins@gmail.com}

}
\keyword{internal}
